+++ app/Controllers/TasksController.js
@@ -1,5 +1,6 @@
 import { ProxyState } from "../AppState.js";
 import { tasksService } from "../Services/TasksService.js";
+import { ListsController } from "./ListsController.js";
 
 
 
@@ -22,15 +23,22 @@ export class TasksController{
     let form = event.target
     let taskData = {
       taskInfo: form.taskName.value,
-      listId: listId
+      listId: listId,
+      
     }
     tasksService.createTask(taskData)
     form.reset()
   }
 
   deleteTask(taskId){
-    tasksService.deleteTask(taskId)
+    let result = window.confirm('Do you want to delete this?')
+    if(result == true){
+      tasksService.deleteTask(taskId)
+    }
 
   }
-
+  
+  setChecked(taskId){
+    tasksService.setChecked(taskId)
+  }
 }
